Move reusable functions to git module
